+++ src/color.rs
@@ -1,4 +1,4 @@
-use core::fmt::{Display, Write as _};
+use core::fmt::Display;
 use std::io::IsTerminal;
 
 use crate::Color;
@@ -47,17 +47,40 @@ pub enum Bg {
     White = 107,
 }
 
-pub type Flags = u8;
-pub struct Flag;
+#[expect(
+    non_camel_case_types,
+    reason = "behaves like a primitive type, so it should be named like a primitive type"
+)]
+pub type flag = u8;
 
 #[expect(non_upper_case_globals, reason = "used as type safe constants")]
-impl Flag {
-    pub const Default: Flags = 0b0000_0000;
-    pub const Bold: Flags = 0b0000_0001;
-    pub const Underline: Flags = 0b0000_0010;
-    pub const NoUnderline: Flags = 0b0000_0100;
-    pub const ReverseText: Flags = 0b0000_1000;
-    pub const PositiveText: Flags = 0b0001_0000;
+pub mod flags {
+    use super::flag;
+
+    pub const Default: flag = 0b0000_0000;
+    pub const Bold: flag = 0b0000_0001;
+    pub const Underline: flag = 0b0000_0010;
+    pub const NoUnderline: flag = 0b0000_0100;
+    pub const ReverseText: flag = 0b0000_1000;
+    pub const PositiveText: flag = 0b0001_0000;
+}
+
+#[expect(
+    non_camel_case_types,
+    reason = "behaves like a primitive type, so it should be named like a primitive type"
+)]
+pub type ansi_code = u8;
+
+#[expect(non_upper_case_globals, reason = "used as type safe constants")]
+pub mod ansi_codes {
+    use super::ansi_code;
+
+    pub const Default: ansi_code = 0;
+    pub const Bold: ansi_code = 1;
+    pub const Underline: ansi_code = 4;
+    pub const NoUnderline: ansi_code = 24;
+    pub const ReverseText: ansi_code = 7;
+    pub const PositiveText: ansi_code = 27;
 }
 
 #[expect(non_upper_case_globals, reason = "it's a function, so it should be named like a function")]
@@ -65,7 +88,7 @@ pub(crate) static mut print: fn(
     &str,
     Fg,
     Bg,
-    Flags,
+    flag,
     &mut core::fmt::Formatter<'_>,
 ) -> core::fmt::Result = print_color;
 
@@ -86,7 +109,7 @@ fn print_no_color(
     text: &str,
     _: Fg,
     _: Bg,
-    _: Flags,
+    _: flag,
     f: &mut core::fmt::Formatter<'_>,
 ) -> core::fmt::Result {
     return text.fmt(f);
@@ -96,43 +119,78 @@ fn print_color(
     text: &str,
     fg: Fg,
     bg: Bg,
-    flags: Flags,
+    flags: flag,
     f: &mut core::fmt::Formatter<'_>,
 ) -> core::fmt::Result {
-    // TODO(stefano): fill a preallocated static buffer instead of creating a new one each time
-    let mut codes = String::with_capacity(24);
+    const CODES_LEN: usize = 24;
+
+    let mut codes_bytes = 0_u64;
 
     if fg != Fg::Default {
-        _ = write!(codes, "{};", fg as u8);
+        codes_bytes |= fg as ansi_code as u64;
+        codes_bytes <<= u8::BITS as u64;
     }
     if bg != Bg::Default {
-        _ = write!(codes, "{};", bg as u8);
+        codes_bytes |= bg as ansi_code as u64;
+        codes_bytes <<= u8::BITS as u64;
     }
-    if flags & Flag::Bold != 0 {
-        codes += "1;";
+    if flags & flags::Bold != 0 {
+        codes_bytes |= ansi_codes::Bold as u64;
+        codes_bytes <<= u8::BITS as u64;
     }
-    if flags & Flag::Underline != 0 {
-        codes += "4;";
+    if flags & flags::Underline != 0 {
+        codes_bytes |= ansi_codes::Underline as u64;
+        codes_bytes <<= u8::BITS as u64;
     }
-    if flags & Flag::NoUnderline != 0 {
-        codes += "24;";
+    if flags & flags::NoUnderline != 0 {
+        codes_bytes |= ansi_codes::NoUnderline as u64;
+        codes_bytes <<= u8::BITS as u64;
     }
-    if flags & Flag::ReverseText != 0 {
-        codes += "7;";
+    if flags & flags::ReverseText != 0 {
+        codes_bytes |= ansi_codes::ReverseText as u64;
+        codes_bytes <<= u8::BITS as u64;
     }
-    if flags & Flag::PositiveText != 0 {
-        codes += "27;";
+    if flags & flags::PositiveText != 0 {
+        codes_bytes |= ansi_codes::PositiveText as u64;
+        codes_bytes <<= u8::BITS as u64;
     }
 
-    return if codes.is_empty() {
-        text.fmt(f)
-    } else {
-        let _last_semicolon = codes.pop();
+    codes_bytes >>= u8::BITS as u64;
+    if codes_bytes == 0 {
+        return text.fmt(f);
+    }
 
-        write!(f, "\x1b[{codes}m")?;
-        text.fmt(f)?;
-        write!(f, "\x1b[0m")
-    };
+    let mut codes: [u8; CODES_LEN] = [b';'; CODES_LEN];
+    let mut codes_end_pointer = codes.as_mut_ptr().wrapping_add(codes.len());
+
+    loop {
+        let mut code = codes_bytes as u8;
+        loop {
+            codes_end_pointer = codes_end_pointer.wrapping_sub(1);
+            unsafe { *codes_end_pointer = (code % 10) + b'0'; }
+            code /= 10;
+            if code == 0 {
+                break;
+            }
+        }
+        codes_end_pointer = codes_end_pointer.wrapping_sub(1);
+
+        codes_bytes >>= u8::BITS as u64;
+        if codes_bytes == 0 {
+            break;
+        }
+    }
+
+    // skipping the last semicolon
+    codes_end_pointer = codes_end_pointer.wrapping_add(1);
+    let codes_len = CODES_LEN - (codes_end_pointer as usize - codes.as_ptr() as usize);
+
+    let codes_slice = unsafe { core::slice::from_raw_parts(codes_end_pointer, codes_len) };
+    let codes_str = unsafe { core::str::from_utf8_unchecked(codes_slice) };
+
+    write!(f, "\x1b[{codes_str}m")?;
+    text.fmt(f)?;
+    return write!(f, "\x1b[0m");
 }
 
 #[derive(Debug, Default)]
@@ -140,7 +198,7 @@ pub struct Colored<Text: AsRef<str>> {
     pub text: Text,
     pub fg: Fg,
     pub bg: Bg,
-    pub flags: Flags,
+    pub flags: flag,
 }
 
 impl<Text: AsRef<str>> Display for Colored<Text> {